+++ packages/replay-data/src/analysis/dependency-graph-shared.ts
