+++ packages/replay-data/src/analysis/specs/analyze-dependencies.ts
