+++ packages/replay-data/src/analysis/specs/execution-point.ts
